pyproject.toml
[tool.poetry]
name = "github-runners-for-repo"
version = "0.1.0"
description = "Zero-install containerized GitHub Actions self-hosted runners for a specific repository"
authors = []
readme = "README.md"
license = "MIT"
packages = [{include = "github_runners_for_repo", from = "."}]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31"
python-dotenv = "^1.0"
docker = "^7.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-cov = "^5.0"
ruff = "^0.4"

[tool.poetry.scripts]
gh-runners = "github_runners_for_repo.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
